+++ src/post.css
@@ -10,7 +10,7 @@ body { font-family: "Pretendard Variable", Pretendard, -apple-system, BlinkMacSy
 article, header, nav, section { display: block; }
 a { text-decoration: none; color: inherit; }
 ul { padding: 0; margin: 0; list-style: none;}
-.container { margin: 0 auto; height: 100%; max-width: 900px; width: 100%; }
+.container { margin: 0 auto; height: 100%; max-width: 900px; width: 100%; padding: 0 40px; }
 
 /* +++++++++++++++++ 헤더 +++++++++++++++++ */
 header {
@@ -92,18 +92,18 @@ h2 {
 .list_area .post_title {
     line-height: 1.4;
     font-size: 18px;
-    width: 80%;
+    width: 74%;
     font-weight: 700;
     word-break: break-word;
 }
 .list_area .post_no {
-    width: 8%;
+    width: 10%;
     font-size: 16px;
     color: #999999;
     text-align: center;
 }
 .list_area .post_date {
-    width: 12%;
+    width: 26%;
     font-size: 16px;
     color: #999999;
     text-align: center;
